Adjust menu bar position
