+++ src/main/resources/view/DarkTheme.css
@@ -193,11 +193,12 @@
 }
 
 .menu-bar {
-    -fx-background-color: derive(#1d1d1d, 20%);
+   /*-fx-background-color: derive(#1d1d1d, 20%);*/
+   -fx-background-color: derive(#1d1d1d, 30%);
 }
 
 .menu-bar .label {
-    -fx-font-size: 14pt;
+    -fx-font-size: 10pt;
     -fx-font-family: "Segoe UI Light";
     -fx-text-fill: white;
     -fx-opacity: 0.9;
